// frontend/src/generated/types/fileResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface FileResponse {
  sysId?: string;
  originalFileNm?: string;
  fileSize?: string;
  filePath?: string;
  ordNo?: number;
  fileExt?: string;
  pdfYn?: string;
}